Fixed a conversion problem in the TileImage.h file.

+++ inc/TileImage.h
@@ -113,9 +113,8 @@ namespace degate
 		{
 			if (!file_exists(_directory)) create_directory(_directory);
 
-			unsigned temp_tile_size = 1 << _tile_width_exp;
-			tiles_number = static_cast<unsigned>(ceil(_width / temp_tile_size)) * static_cast<unsigned>(ceil(
-				_height / temp_tile_size));
+			double temp_tile_size = 1 << _tile_width_exp;
+			tiles_number = static_cast<unsigned>(ceil(static_cast<double>(_width) / temp_tile_size)) * static_cast<unsigned>(ceil(static_cast<double>(_height) / temp_tile_size));
 		}
 
 		/**
